+++ kitsune/Cargo.toml
@@ -51,6 +51,7 @@ http = "0.2.9"
 hyper = { version = "0.14.27", features = ["deprecated"] }
 iso8601-timestamp = "0.2.11"
 kitsune-cache = { path = "../kitsune-cache" }
+kitsune-common = { path = "../kitsune-common" }
 kitsune-db = { path = "../kitsune-db" }
 kitsune-embed = { path = "../kitsune-embed" }
 kitsune-http-client = { path = "../kitsune-http-client" }
