+++ kitsune/src/http/page.rs
@@ -1,12 +1,12 @@
 use crate::{
     error::{Error, Result},
     state::Zustand,
-    try_join,
 };
 use askama::Template;
 use diesel::{BelongingToDsl, QueryDsl, SelectableHelper};
 use diesel_async::RunQueryDsl;
 use futures_util::{future::OptionFuture, TryStreamExt};
+use kitsune_common::try_join;
 use kitsune_db::{
     model::{
         account::Account,
