+++ kitsune/src/mapping/activitypub/activity.rs
@@ -1,9 +1,10 @@
 use super::IntoObject;
-use crate::{error::Result, state::Zustand, try_join};
+use crate::{error::Result, state::Zustand};
 use async_trait::async_trait;
 use diesel::QueryDsl;
 use diesel_async::RunQueryDsl;
 use iso8601_timestamp::Timestamp;
+use kitsune_common::try_join;
 use kitsune_db::{
     model::{account::Account, favourite::Favourite, follower::Follow, post::Post},
     schema::{accounts, posts},
