+++ kitsune/src/mapping/mastodon/sealed.rs
@@ -1,7 +1,6 @@
 use crate::{
     error::{Error, Result},
     service::{attachment::AttachmentService, url::UrlService},
-    try_join,
 };
 use async_trait::async_trait;
 use diesel::{
@@ -10,6 +9,7 @@ use diesel::{
 };
 use diesel_async::RunQueryDsl;
 use futures_util::{future::OptionFuture, FutureExt, TryFutureExt, TryStreamExt};
+use kitsune_common::try_join;
 use kitsune_db::{
     model::{
         account::Account as DbAccount,
